requirements.txt
absl-py==0.9.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.5.0
cycler==0.10.0
Cython==0.29.21
dateparser==0.7.6
DateTime==4.3
ExifRead==2.3.1
Flask==1.1.2
future==0.18.2
fvcore==0.1.1.post20200716
google-api-core==1.22.1
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mock==4.0.2
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.4.0.40
Pillow==7.2.0
portalocker==2.0.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.1
pydot==1.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
Shapely==1.7.0
six==1.15.0
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.2
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
yacs==0.1.8
